HealthBuddy/requirements.txt
streamlit
pandas
numpy
matplotlib
pillow
pytesseract
PyPDF2
google-generativeai
opencv-python-headless
python-dotenv
